feat: :sparkles: implement useUploadFiles() hook
